refactor(LottosResults) : LottoResults에서 LottosResults로 클래스명 변경
